feat: [pipe-20761]: migrate pr list-comments
