--- apps/gitness/src/components/RepoWrapper.tsx
@@ -1,11 +1,0 @@
-import { RepoLayout } from '@harnessio/ui/views'
-
-const RepoWrapper = () => {
-  return (
-    <>
-      <RepoLayout />
-    </>
-  )
-}
-
-export default RepoWrapper
