+++ packages/ui/package.json
@@ -120,8 +120,7 @@
     "vite-plugin-dts": "^4.3.0",
     "vite-plugin-svgr": "^4.3.0",
     "vite-tsconfig-paths": "^5.1.2",
-    "vitest": "^2.1.5",
-    "classnames": "^2.5.1"
+    "vitest": "^2.1.5"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
